// src/app/models/UsersModel.ts
export interface UsersModel1
{
    id: number | any; 
    name: string | any;
    email: string | any;
    address: string | any;
    phone: string | any;
    birthday: Date | any;
}
export interface UsersModel2
{
    id: number | any; 
    name: string | any;
    email: string | any;
    address: string | any;
    phone: string | any;
    birthday: Date | any;
    age: number | any;
}